chore(server): move config, middleware, models, utils under src/ and fix all import/test paths for rootDir consistency
