+++ server/src/controllers/gameController.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { supabase } from '../../config/database';
+import { supabase } from '../config/database';
 
 export const getGames = (req: Request, res: Response) => {
   res.status(200).json([{ id: 1, title: 'Game 1' }, { id: 2, title: 'Game 2' }]);
