+++ server/src/utils/caseMapping.ts
