+++ server/src/utils/supabaseSchema.sql
